refactor(application): update use cases to use entities instead of models
