+++ src/application/usecases/authenticate-user/authenticate-user.usecase.test.ts
@@ -3,8 +3,8 @@ import { PasswordHasherStub } from '@/infra/adapters/security/stubs/password-has
 import { InMemoryRefreshTokensRepository } from '@/infra/persistence/repositories/in-memory/in-memory-refresh-tokens.repository'
 import { InMemoryUsersRepository } from '@/infra/persistence/repositories/in-memory/in-memory-users.repository'
 import type { PasswordHasher } from '@/infra/adapters/security/ports/password-hasher'
-import { makeUser } from '@/util/factories/domain/make-user'
 import { AuthenticateUserUseCase } from './authenticate-user.usecase'
+import { makeUser } from '@/shared/util/factories/domain/make-user'
 
 vi.mock('@/lib/env', () => ({
   env: {
