+++ src/application/usecases/authenticate-user/authenticate-user.usecase.ts
@@ -3,9 +3,9 @@ import type { RefreshTokensRepository } from '@/application/repositories/refresh
 import type { UsersRepository } from '@/application/repositories/users.repository'
 import { JWTService } from '@/infra/auth/jwt/jwt-service'
 import type { PasswordHasher } from '@/infra/adapters/security/ports/password-hasher'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { RefreshToken } from '@/domain/entities/refresh-token/refresh-token.entity'
 import { InvalidPasswordError } from './errors/invalid-password.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 export type AuthenticateUserRequest = {
   email: string
