+++ src/application/usecases/choose-question-best-answer/choose-question-best-answer.usecase.ts
@@ -1,9 +1,9 @@
 import type { UseCase } from '@/core/application/use-case'
 import type { AnswersRepository } from '@/application/repositories/answers.repository'
 import type { QuestionsRepository } from '@/application/repositories/questions.repository'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import { Question } from '@/domain/entities/question/question.entity'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 export type ChooseQuestionBestAnswerRequest = {
   authorId: string
