+++ src/application/usecases/delete-comment/delete-comment.usecase.test.ts
@@ -1,10 +1,10 @@
 import { InMemoryAnswersRepository } from '@/infra/persistence/repositories/in-memory/in-memory-answers.repository'
 import { InMemoryCommentsRepository } from '@/infra/persistence/repositories/in-memory/in-memory-comments.repository'
 import { InMemoryQuestionsRepository } from '@/infra/persistence/repositories/in-memory/in-memory-questions.repository'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeComment } from '@/util/factories/domain/make-comment'
 import { DeleteCommentUseCase } from './delete-comment.usecase'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeComment } from '@/shared/util/factories/domain/make-comment'
 
 describe('DeleteCommentUseCase', () => {
   let sut: DeleteCommentUseCase
