+++ src/application/usecases/delete-question/delete-question.usecase.test.ts
@@ -1,9 +1,9 @@
 import type { QuestionsRepository } from '@/application/repositories/questions.repository'
 import { InMemoryQuestionsRepository } from '@/infra/persistence/repositories/in-memory/in-memory-questions.repository'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeQuestion } from '@/util/factories/domain/make-question'
 import { DeleteQuestionUseCase } from './delete-question.usecase'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeQuestion } from '@/shared/util/factories/domain/make-question'
 
 describe('DeleteQuestionUseCase', () => {
   let sut: DeleteQuestionUseCase
