+++ src/application/usecases/fetch-answer-comments/fetch-answer-comments.usecase.test.ts
@@ -1,7 +1,7 @@
 import type { AnswerCommentsRepository } from '@/application/repositories/answer-comments.repository'
 import { InMemoryAnswerCommentsRepository } from '@/infra/persistence/repositories/in-memory/in-memory-answer-comments.repository'
-import { makeAnswerComment } from '@/util/factories/domain/make-answer-comment'
 import { FetchAnswerCommentsUseCase } from './fetch-answer-comments.usecase'
+import { makeAnswerComment } from '@/shared/util/factories/domain/make-answer-comment'
 
 describe('FetchAnswerCommentsUseCase', () => {
   let sut: FetchAnswerCommentsUseCase
