+++ src/application/usecases/get-user-by-email/get-user-by-email.usecase.ts
@@ -1,7 +1,7 @@
 import type { UseCase } from '@/core/application/use-case'
 import type { UsersRepository } from '@/application/repositories/users.repository'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import type { User } from '@/domain/entities/user/user.entity'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 type GetUserByEmailUseCaseRequest = {
   email: string
