+++ src/application/usecases/update-account/update-account.usecase.test.ts
@@ -3,9 +3,9 @@ import { PasswordHasherStub } from '@/infra/adapters/security/stubs/password-has
 import {
   InMemoryUsersRepository,
 } from '@/infra/persistence/repositories/in-memory/in-memory-users.repository'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeUser } from '@/util/factories/domain/make-user'
 import { UpdateAccountUseCase } from './update-account.usecase'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeUser } from '@/shared/util/factories/domain/make-user'
 
 describe('UpdateAccountUseCase', () => {
   let sut: UpdateAccountUseCase
