+++ src/application/usecases/update-answer/update-answer.usecase.test.ts
@@ -2,9 +2,9 @@ import type { AnswersRepository } from '@/application/repositories/answers.repos
 import {
   InMemoryAnswersRepository,
 } from '@/infra/persistence/repositories/in-memory/in-memory-answers.repository'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeAnswer } from '@/util/factories/domain/make-answer'
 import { UpdateAccountUseCase } from './update-answer.usecase'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeAnswer } from '@/shared/util/factories/domain/make-answer'
 
 describe('UpdateAccountUseCase', () => {
   let sut: UpdateAccountUseCase
