+++ src/application/usecases/update-comment/update-comment.usecase.test.ts
@@ -1,8 +1,8 @@
 import { InMemoryCommentsRepository } from '@/infra/persistence/repositories/in-memory/in-memory-comments.repository'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeComment } from '@/util/factories/domain/make-comment'
 import { UpdateCommentUseCase } from './update-comment.usecase'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeComment } from '@/shared/util/factories/domain/make-comment'
 
 describe('UpdateCommentUseCase', () => {
   let sut: UpdateCommentUseCase
