+++ src/application/usecases/update-comment/update-comment.usecase.ts
@@ -1,8 +1,8 @@
 import type { UseCase } from '@/core/application/use-case'
 import type { CommentsRepository } from '@/application/repositories/comments.repository'
-import { NotAuthorError } from '@/application/errors/not-author.error'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
 import type { Comment } from '@/domain/entities/comment/comment.entity'
+import { NotAuthorError } from '@/shared/application/errors/not-author.error'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
 
 export type UpdateCommentRequest = {
   commentId: string
