+++ src/application/usecases/update-question/update-question.usecase.test.ts
@@ -2,9 +2,9 @@ import type { QuestionsRepository } from '@/application/repositories/questions.r
 import {
   InMemoryQuestionsRepository,
 } from '@/infra/persistence/repositories/in-memory/in-memory-questions.repository'
-import { ResourceNotFoundError } from '@/application/errors/resource-not-found.error'
-import { makeQuestion } from '@/util/factories/domain/make-question'
 import { UpdateQuestionUseCase } from './update-question.usecase'
+import { ResourceNotFoundError } from '@/shared/application/errors/resource-not-found.error'
+import { makeQuestion } from '@/shared/util/factories/domain/make-question'
 
 describe('UpdateQuestionUseCase', () => {
   let sut: UpdateQuestionUseCase
